refactor(back): reorganiza modulo identity de DDD em camadas para package-by-feature
